MotorController new: LUT DMA CAN update: PID topdesign note:har ændret i Generated_Source så ik' fjern mappen!!!

+++ Software/MotorControllerPSoC/MotorController/MotorController.cydsn/Pedal.h
@@ -11,7 +11,7 @@
 */
 #include <project.h>
 
-const uint16 *Pedal_init();
+const uint8 *Pedal_init();
 
 void Pedal_update();
